Add Bower package and config
OMG what's old is… still old. This addition is a gnarly hack so that we
can use an old old old version of the Octicons package (v2.0.1) that
was released for Bower but not npm.

In other words, the old package has a `bower.json` config, but not a
`package.json` meaning it can't be installed via `npm install`.

So… we install Bower locally and add a postinstall task to install a
really old package from GitHub into `node_modules`.

Oi.